To fix "Unable to get property 'replace' of undefined - only in IE9".
To fix "Unable to get property 'replace' of undefined - only in IE9".

I seem to always get this issue with angular in IE9 (but never in IE10+ or chrome etc). I suspect it's due to how strict IE9 is with illegal markups and error handling in general. But even though my code passed all HTML validation tools, this error still occurred.